Refactor ResultsListItem
Move the divs showing the bolded query matches into matched-string.js, so we can reuse them in the tester.
Add get-tabs.js to handle adding URLs to the tabs.
Rename display-url.js to add-urls.js, and change it to add the displayURL and faviconURL, so we can do that just once instead of in ResultsListItem.render().