Search across 2000 history items and fix search box cursor
Loop through multiple history search calls to build up a list of 2000 items.
Fix bug where moving the insertion point from the end of the query and then typing moved it back to the end.
Added input.js component that wraps an <input> to ignore changes when it's not focused.