Virtualize <ResultsList> with react-virtualized.js
Don't limit the matches to 10 and let the user scroll through them.
Reduce the MinScore to show more matches, while dropping nearly zero items even when the matches get to 3.
Switch favicon to its own element so that we can fade it if The Great Suspender has forgotten its own faded icon.
Discount the scores of strings > 150 chars.
Alt-click a result to copy info about it to the clipboard in dev mode.
Remove chrome:// links from index.md, since they don't seem to work on a webpage.
Map different names for React and ReactDOM in the RJS config to support react-virtualized.js.
Added react-virtualized.js and rv-styles.css.
Added copy-to-clipboard.js.