Refactor onQueryChange() to buffer chars while bookmarks/history are loading
If the user quickly types /h or /b and then a query, the items could still be loading via the promise while the user's still typing.  So the characters would appear on top of the placeholder text.
Instead, buffer the characters in onKeyDown() and force the input to not change.  Then restore the full query when the promise resolves, and then show the matching items.