Rework keyboard event handling
Add shortcut-manager.js and handle-keys.js to dispatch key events to the right handlers.
Move keyboard handling to handle-keys.js.
Move adding the unsuspendURL to addURLs() instead of getTabs() so we can call it when moving tabs, since the tab reference changes and we may want to unsuspend the tab.
Add clearQuery() method.
Remove the bookmarks and history regexes.
Upped the minimum Chrome version to 45 because we're using Object.assign().