Use mutex when saving adding and removing tabs
When a tab is closed, the removed and activated events fire quickly, and the data saved by addTab() was usually out of date.
Add support for pressing space to navigate to next item.
Update ShortcutManager to support spaces.
Add onStartup event handler, outside the main require() in background.js.
Add version to the stored data.
Add updateRecents() method to fix the tab and window IDs for the stored recents after the browser restarts.
Reduce recent boosts.