Fix Tracker to work if analytics.js loads after it's created
Once the local ga was set, it was never updated, so if the GA file loaded slowly, events would never fire.
Add timeout so that if the GA code never loads, the tracker will clear the queue and disable itself, so the queue doesn't grow forever.
Refactor calling the ga object in Tracker.
Change Tracker constructor to take an object instead of params.
Change default tracker name to t0.