Load error-handler.js before other scripts
Retry if global require() is not available yet so we can load before RequireJS.
Use a single string for logging errors, since the extensions page only lists the first argument to console.error().
Check runtime.lastError in init.js to avoid errors on extension page.
Make global DEBUG a var instead of const, so it can be changed via the console.