Fix issue #5
When focusing a tab, focus the window before focusing the tab, which seems to avoid focus issues on macOS.
In popups, call promisified update() methods in a chain instead of the non-callback versions.
Remove code for opening options page after upgrade to 1.1.0.
Check for error.lineno when tracking an exception, so we don't send undefineds.