Add option to show icon for tabs in other windows
Update storageVersion to 7.
Calculate the storage version by incrementing the highest updater key, so that we don't have to explicitly change the version after adding an updater.
Add ability to show a subtitle to checkboxes and radio buttons.
Use functional components for the icons on the Options page, instead of JSX objects.
Remove UpgradeMessage from Options, which we're no longer using.
Use destructuring to avoid having to use Control. to access the control components.
Change differentWindow to otherWindow, to match the setting name.