Update results list when a tab is closed
Refactor componentWillMount() to be much simpler.
Add initTabs() method to app.
Add a tab.onRemoved handler to re-init the list.
Check that the item has a tab ID before trying to close it.
Renamed get-tabs.js to init-tabs.js.
Add includeClosedTabs param to recentTabs.getAll(), so that that module doesn't need to check the settings.
Clean up the createStorage() params.