Score items in loadPromisedItems()
Just sort the history items when getting a /h command in getMatchingItems(), instead of scoring them, too.
Add a reload param to loadPromisedItems() to force it to replace the existing promise.
Update releases/index.md to list updating the results if a tab is closed while the menu is open.