Move addURLs() to recentTabs.getAll()
When deleting a closed tab, also remove it from the recents list, so that if the user just scrolled down to the closed tabs section and deleted one, it would disappear.
Filter out unsuspended closed tabs that have the same URL as suspended open tabs.
Fix renamed initTabs() in app.js.
Return the item from addURLs().