Add lastSeenOptionsVersion to storage
Use it to show a badge on the options button when storage.version is greater than it, and show NEW badges next to new settings in options.
Access storage from the popup instead of just settings, since it needs the lastSeenOptionsVersion and storage version.
Add newSettingsAvailable to popup state.
Refactor options button markup into OptionsButton.
Don't pass separate shortcuts and chrome props to OptionsApp, and let it just get those from the settings prop.
Add NewSetting component to wrap new settings that we add to the options page going forward, so that a NEW badge is shown next to them if the user hasn't opened options recently.
Add version getter to storage.
Change updaters to arrow functions.
Enable passing storage data into settings.get() so that if we already have it, like in the popup, settings doesn't have to get it again.
Update popup.html with the latest rendered JSX.