Support searching via pinyin
Create PinyinQuickScoreArray() to use the pinyin fields when scoreItems() is called with a paramter.
Wrap pinyin function to always use STYLE_NORMAL and collapse the array of results into a string.
Add pinyinTitle and pinyinDisplayURL in initTabs(), based on a param.
Add pinyinTitle and pinyinDisplayURL to the item tooltips in dev mode.
Always browserify and clean rjs before other steps in grunt build.