Use import() to load pinyin on demand
Works fine when running from source, but r.js and esprima choke on the import() keyword when building, so this still needs to get fixed.
Don't include pinyin.js in the main.js bundle.
Add grunt task to prepend the webpack output of pinyin with `export ` so it can imported with import().
Add optional heteronym param to pinyin().