Move popup off-screen instead of minimizing it
It's much faster to move the window than minimize it, especially on macOS.
Update the recents list in the background when tabs are activated, so it doesn't have to re-render when focused.
Add visible flag to the popup app so it can behave differently in the background vs. foreground.
Send tabActivated messages from background to the popup when tabs are focused.
Remove focus from the popup if it's closed by pressing esc.
Clear the search query when the popup loses focus.
Don't add the popup window to the recent tabs list.
Don't specify the width/height when the background moves the popup back on screen, to avoid window redrawing.
Make sure the popup is on screen when centering it over the active window.