Use windows.create() to create the popup window
Add isPopup prop, instead of checking menubar.visible, since it's true in popup windows created by the API.
Call closeWindow() before focusing a tab.   Only call it in a timeout if it's not a popup window.
Move popup command to second position in the shortcuts list.
Add closePopupWindow() function.
Add constants for the manifest command IDs and off-screen positions.
Close the popup window if the port is closed.