Support persisting the popup as a tab, instead of an off-screen window
Add type to popupWindow, and use it to either create the popup as a window or a tab and to determine how it's shown/hidden.
In tab mode, hide the popup as a tab in the last normal window.
Pass type and focusSearch params on the URL when opening the popup, so it knows how to behave on first launch.
Don't log errors in popupWindow.close().
Close all the open tabs/windows when the background is reloaded.
Don't invert the icon when hitting the next tab command while the user isn't navigating.
Add openedForSearch member to the popup, so it can set the selected index correctly depending on how the window was opened.  Avoids an extra render.
Add focusSearch prop to popup.
Log event.reason.message in error-handler.js, in case it's a getter.