Don't focus item under mouse when popup is opened
Reset mouseMoveCount in the list items when the item is changed or the list is hidden, so that the item won't get selected if it's opened under the mouse.
Focus the searchbox when the popup window is shown, in case something else had focus the last time it was shown.
Filter out incognito windows from getLastWindow().
Catch errors in popupWindow.show() and hide().
ES6ify results-list.js.