Hide the popup behind other windows if on Mac or pixel ratio !== 1.0
Await for the changed tab or window in openItem() so we can pass it to popupWindow.hide(), which can hide the popup behind the focused window, if needed.
Listen for resolution changes in the popup and close it if it changes.
Add popup size adjustments to storage so they can be reused the next time the extension is reloaded.
Fix the size adjustments so they actually update correctly when needed.