Add navigatingRecents mode to popup app
Move handling of focusing tab when it's selected to onMessage(), instead of modifySelected().
Fix handling of opening the selected item on keyup in normal popup mode.
Add navigateRecentsWithPopup var in background.
When navigateRecentsWithPopup is true, tell the popup to modify the selection, select the tab, and re-show the popup after.
Fix centering alignment of popup.