Refactor popup window controls into popup-section.js
Show the pros/cons to the right of the popup options on hover.
Use colors for the pro/con labels.
Pass rest of props to controls so we can add styles, event handlers, etc.
When the popup is blurred by another window getting focused, query for that window to hide behind instead of what was active when the popup was shown, since it may have changed.