Reopen popup if it detects inner and outer width are the same
Send message to background to close and reopen the popup.
Only close the popup when the port is closed if it's still open.
Set the tab and window IDs to 0 in popupWindow.close() before tabs.remove() is called, so that if something checks isOpen before the tab is removed, it won't find the tab that's in the process of being closed.
Store the popup's ID so that it can ignore the onTabRemoved event that happens when the popup itself is closed.