Add option to show popup while navigating recents
Add checkbox for the option in Popup window section.
Default the option to false.
Handle changes to the option in background.js and app-container.js.
Add reopenWindow() method to the popup app.
Add link to add a review to About section.
Refactor .shortcut styles so <Shortcut> can also be used in labels, outside of a <ShortcutPicker>.
Log when the popup borders collapse.