Add more shortcuts for vimium-like navigation
This adds the ability to set custom shortcuts for up/down in the picker,
so you can do ctrl-p and ctrl-n as you do with the omnibar and the
vomnibar in vimium. It also adds the ability to set custom shortcut
behavior, so ctrl-[ escapes as it does in vim.