Score tabs with empty query before filtering them
If limit search is on but limit recents is not, some recents may not appear in the filtered tabs list, and therefore won't get the scores hash and other keys set up on the first scoreItems() call.  So score all tabs before filtering them, so all the recents get set up, too.
Add "browser" to the limit options.