Link the limit recents and limit search options
Check recents when the limit search to current window option is checked.
Uncheck search if recents is unchecked.  This way, we don't need to enable/disable the options.
Don't call scoreItems() inside loadTabs(), since now the recents list is guaranteed to be a subset of the filtered tabs, since the limit search results option is linked to the limit recents option.