Allow menu height to grow in Firefox
Add support for the search box cancel button in FF by rendering it with React, since FF doesn't render the pseudo element.
Fix flash of scrollbar when menu is initially shown.
Use scrollbar-width and -color to style the scrollbar in FF, since it doesn't support as many options as Webkit.
Don't show a favicon for internal FF pages, since it generates console errors.
Add a .firefox class to the root.
Add IsFirefox constant.