Prevent popup window from going off-screen when centering on a window
Control whether an item gets selected on a mouse move in the results list, by waiting for a timer to fire after the window is shown.  That way, an item doesn't get selected if it's rendered under the mouse, and the logic doesn't have live on every item.
Update prop names passed from the popup to the items.
Default app.visible to false, so the drop-down menu sees a change in that prop on the render list.