Make popup work when it's opened for the first time to navigate recents
Have the popup app detect that it's being opened to navigate recents and then change the selection when it mounts.
Don't show the popup if the switch to next command is pressed and it's not currently open.
Pass flags from background to popup as `props` JSON string.
In results list, only call setSelectedIndex() if the index has changed.