Add "Show popup window" subsection to top of "Popup window"
List the show popup window shortcut and a description.
Move HidePopup animation above the Hide popup subsection.
Control which demo auto starts by passing a prop from popup-section.js.
Rename demo/Shortcut to AnimShortcut, so it's different than the one in options/.
Use span instead of div in shortcut.js to avoid warning when using it in a p.