Use react-router for client-side routing
Wrap the app in a HashRouter, since the options.html page isn't served by a server.
Convert OptionsApp into a functional component so it can use the router hooks.
Support hashed search params on the location.
Create withSearchParams() HOC to supply useSearchParams() to a class component.
Wrap OptionsAppContainer with the search params and the HashRouter so it can access the search params.
Move About section into about-section.js.
Move most props to context.
In app-container.js, add the state to the context, so it doesn't have to be passed down as props.
Get lastSeenOptionsVersion from context in NewSetting.
Clean up unused props.