Scale the windows via a CSS variable instead of a context
Scale the popup window contents so they look right on a hi res monitor.
Adjust bounds returned from getWindowBounds() to account for negative locations when the window is on a monitor to the left of the main one.
Make the titles in the popup at least 2px thick.
Fix the position of the popup navigating recents when the screen starts at a negative value.
Make the tab name bold.
Remove unnecessary wrapper and simplify prop handling in Rect.js.
Remove DemoContext.js.