Emit events when popup window changes
Have the background listen for popup hide and close events, so that it can reset the navigatingRecents flags.  This fixes a bug where it would stop tracking tab activations if the popup was in navigate recents mode and got closed somehow.
Add popup-emitter.js.
Change min Chrome version to 65, since we need the EventTarget class.