Emit events when popup window changes state
Have the background listen for popup hide and close events, so that it can reset the navigatingRecents flags.  This fixes a bug where it would stop tracking tab activations if the popup was in navigate recents mode and got closed somehow.
Add popup-emitter.js.
Change min Chrome version to 65, since we need the EventTarget class.
Support Firefox when using popups.
Don't include setSelfAsOpener in popup options in FF, since it's not supported.
Also don't try to use chrome.system if it doesn't exist, which it doesn't in FF.
Refactor calcPosition() calls.
Rename calcPosition() to calcBounds().