Support searching multiple tokens separated by spaces
Update settings version to 13.
Add "all of the above" option to the space shortcut setting, and change it to that when upgrading the version.
Add the control-group class to a NewSetting when it wraps one, so the spacing is consistent.
Use number literals for the version updaters in quickey-storage.js.
Remove logging from getMatchingItems.
Don't clear selection when space is pressed.
Add getSelection() method to SearchBox.
Add all Latin chars, including accented ones, to search.js.
Use total query length and average token length, not just token count, to switch to using simpleScore().
Pass empty array instead of empty string to scoreItems() in loadPromisedItems.
Don't normalize whitespace in init-tabs.js, since we break the query on spaces, so the user can't match on spaces.
Use same scoring for URLs as for titles.
Use alt-shift-click to copy info about the item in dev mode.
Pass tokens array to scoreArray() instead of a string.
Adjust the min score values based on the number of tokens.
Clean up score-items.js.
Allow tokens to match any of the keys.
Only count an item as a match if all of the tokens match at least one the item's keys.
Fix typing a space after starting a command.
In MapCache, just pass the defaultEntries param directly to the Map when creating it.
Move the selection down even when a space is inserted.
Refactor scoreArray() and use for of loop.
Disable hover select when the results list array changes.  This prevents a selection under the mouse when the popup height changes on a non-integer-scaled screen.
When resizing height of popup, use the saved width instead of the current outerWidth.
Ignore small window resize events.  These seem to happen when the devicePixelRatio is not an integer, and can cause repeated resize events because the window doesn't get exactly resized to the requested value.
Use a NUL char for replacement.
Allow space to insert one space at the end of the query and then handle it by moving the selection down.
Fix recency boosting of the query scores.
Changed the default for the floor param to rnd().
Add ctrl-space as a shortcut for selection.
Refactor caching into MapCache.js.
Have unlimited caching of single character regexes.
Use a case-insensitive regex to match the query.  This avoids having to lowercase the query and search string on every loop.