Handle having different modifier keys for the Chrome shortcuts
So the modifier up event will correctly activate the selected item, depending on which shortcut was pressed to open the popup.
Add updateMRUModifier() method.
Handle activation events for tabs in another profile not including the tabId, which was throwing an exception.
Change openPopup to navigatingRecents as a flag in the popup.
In getShortcuts(), add detailed info for all the Chrome shortcuts, not just the one for opening the menu.
Add OpenMenuCommand to constants.