Track which thread last saved to storage and reload as necessary
Have getAll() in storage get fresh data when necessary.
Store a lastSavedFrom key to track the last source of the data.
Check tabId of activated event to the popupWindow.tabID, since checking the saved windowID didn't always work (because it was getting stored in a tab?).
Change min version to 115.
Fix spaces in app.js.