Send messages to the background to get/set storage
Essentially wrap the function passed to storage.get()/set() with calls to the background to first lock the storage mutex, then get the data, pass it to the task in the popup, then send the results back to the background to store.
Listen for get, set, or done messages in the storage module.
Add createStorageClient() to storage.js.
Use Promise.resolve() in returnData().
Convert Mutex to an ES class.