Use a separate service worker to capture events before rest of code is loaded
Add cached event handlers in sw.js for all events used in background.js, then re-dispatch them from the end of background.js.
Use importScripts() to import background.js.
Set DEBUG in background.js and main.js by looking at installType instead of using the defines.js file, which seemed to not get loaded sometimes.
Switch to using `chrome` directly instead of importing cp.js in background.
If the popup window is open when the background loads, reconnect the port.
Always reconnect the channel in ipc.js if the channel name matches, without having to use regexes.
Update the background port used in the popup if the background reconnects it.