Fix too many files getting into the bundle
Comment out an if branch in ga4mp.js that included a call to require(), which completely messed up the webpack bundling and caused everything in lib/ to be bundled.
Put the error queuing into a separate file, so it doesn't pull in the tracker files, and create a separate error-queue entry point.
Import List from a subpath in react-virtualized so it gets tree-shook correctly.
Move setting DEBUG into error-handler.js.
Make npm run build do a production build.
Tweak the welcome message margins.