Fix a timing issue with combining navigate to previous and toggle commands
Fix #123.
Make sure calls to navigate to the previous/next tab follow any pending toggles, and vice versa.
Reset the toolbar icon if the toggle command is executed during the navigation cooldown.
Enable a caller to wait for the next execution of a debounced function and trigger it immediately.
Add toolbar.setNormalIcon() method.
Update edge-store-description.txt to match the Chrome store text.