Refactor event handler creation so they can be easily removed
Add removeListener() method to MessageTarget.
Add onbeforeunload handler when running in the popup to remove chrome event handlers when the popup gets closed, in an attempt to try to get them working again in the service worker.
Wrap importScripts() in a try/catch in sw.js, as recommended in the Google group.